IMAP-66 SELECT should UIDNEXT as per RFC3501 https://issues.apache.org/jira/browse/IMAP-66
git-svn-id: https://svn.apache.org/repos/asf/james/imap/trunk@773250 13f79535-47bb-0310-9956-ffa450edef68

+++ api/src/main/java/org/apache/james/imap/api/message/response/StatusResponse.java
@@ -254,7 +254,7 @@ public interface StatusResponse extends ImapResponseMessage {
          *            positive non-zero integer
          * @return <code>ResponseCode</code>, not null
          */
-        public static final ResponseCode uidNext(int uid) {
+        public static final ResponseCode uidNext(long uid) {
             return new ResponseCode("UIDNEXT", uid);
         }
